+++ apps/client/src/app/modules/tooltip/xivapi-tooltip/xivapi-item-tooltip.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, Input, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Input, OnInit } from '@angular/core';
 import { LocalizedDataService } from '../../../core/data/localized-data.service';
 import { LazyDataService } from '../../../core/data/lazy-data.service';
 
@@ -10,7 +10,8 @@ import { LazyDataService } from '../../../core/data/lazy-data.service';
 })
 export class XivapiItemTooltipComponent implements OnInit {
 
-  constructor(private l12n: LocalizedDataService, private lazyData: LazyDataService) {
+  constructor(private l12n: LocalizedDataService, private lazyData: LazyDataService,
+              private cd: ChangeDetectorRef) {
   }
 
   @Input() item: any;
@@ -25,6 +26,7 @@ export class XivapiItemTooltipComponent implements OnInit {
   public patch: any;
 
   ngOnInit(): void {
+    this.cd.reattach();
     if (this.item === undefined) {
       return;
     }
@@ -118,6 +120,8 @@ export class XivapiItemTooltipComponent implements OnInit {
     if (this.item.ItemSpecialBonus) {
       this.item.SetBonuses = this.lazyData.data.itemSetBonuses[this.item.ID]?.bonuses;
     }
+
+    this.cd.detectChanges();
   }