+++ apps/client/src/environments/patch-notes.ts
@@ -1,11 +1,21 @@
 export const patchNotes = `### Bug Fixes
 
-* **fishing-reporter-overlay:** fixed flickering with patience ending after you threw the line.
-* **leveling-equipment:** fixed an error that happened with crafting filter enabled.
-* **seo:** fixed opensearch integration.
+* **alarms:** complete revamp of the "create alarm in group" system.
+* **alarms:** fixed fishing alarms not being accurate with weather.
+* **fishing-reporter:** fixed random items being detected as bait.
+* **fishing-reporter-overlay:** swapped current and previous weather.
+* **list:** fixed an issue with some items not being added properly to a list.
+* **log-tracker:** fixed an issue with hide completed tabs not selecting another tab properly.
+* **simulator:** fixed a bug with duration per macro not refreshing properly.
 
 
 ### Features
 
-* **data:** update for cn patch-5.41.
-* **simulator:** support for CN 5.41.`;
+* **alarms:** folders names are now shown below alarms in folder addition popup.
+* **db:** support for korean v5.4 update.
+* **simulator:** support for korean 5.4.
+
+
+### Performance Improvements
+
+* **list:** shorter loading time and overall list performance.`;